// 19_MOVIESLIB/movies_lib/src/types/ISimpleMovie.ts
export interface ISimpleMovie {
  id: number;
  title: string;
  adult: boolean;
  backdropPath: string;
  genreIds: number[];
  originalLanguage: string;
  originalTitle: string;
  overview: string;
  popularity: number;
  posterPath: string;
  releaseDate: string;
  video: boolean;
  voteAverage: number;
  voteCount: number;
}